gems/gup.sh: enhance 'continue' workflow

+++ gems/gup.sh
@@ -204,12 +204,14 @@ ask () {
 grep -q "^$PRE" <<< "$PKG" || die "Couldn't autodetect package name: '$PKG'"
 
 # set remote
-bash -c "git fetch '$ORG' $SIL" || die 'Failed to git fetch $ORG'
-bash -c "git fetch '$ME' $SIL" || {
-  bash -c "$FRK '$PKG' $SIL"
-  git remote -v | grep -q "^$ME" \
-    || git remote add "$ME" "git+ssh://$ME@pkgs.fedoraproject.org/forks/$ME/rpms/${PKG}.git"
-  bash -c "git fetch '$ME' $SIL" || warn "Failed to fetch" "$ME"
+[[ -n "$CON" ]] || {
+  bash -c "git fetch '$ORG' $SIL" || die "Failed to git fetch" "$ORG"
+  bash -c "git fetch '$ME' $SIL" || {
+    bash -c "$FRK '$PKG' $SIL"
+    git remote -v | grep -q "^$ME" \
+      || git remote add "$ME" "git+ssh://$ME@pkgs.fedoraproject.org/forks/$ME/rpms/${PKG}.git"
+    bash -c "git fetch '$ME' $SIL" || warn "Failed to fetch" "$ME"
+  }
 }
 
 # status
@@ -270,7 +272,7 @@ nam="`cut -d'-' -f2- <<< "$nam"`"
 }
 
 # new
-rm *.gem ||:
+[[ -n "$CON" ]] || rm *.gem ||:
 bash -c "gem fetch $PRF '$nam' '$ver' $SIL" || die "gem fetch $prf failed"
 
 f="$(basename -s '.gem' "`ls *.gem | tail -n -1`")"
@@ -294,11 +296,13 @@ xv="`rev <<< "$f" | cut -d'-' -f1 | rev`"
   prever=".`rev <<< "$ver" | cut -d'.' -f1 | rev`"
   ver="`rev <<< "$ver" | cut -d'.' -f2- | rev`"
 
-  grep -qE '^[#%]*%global prerelease' "$X" && {
-    sed -i "s/^[#%]*\(%global prerelease\).*$/\1 $prever/" "$X"
-    :
-  } || {
-    sed -i "/^\s*Name: / i %global prerelease $prever\n" "$X"
+  [[ -n "$CON" ]] || {
+    grep -qE '^[#%]*%global prerelease' "$X" && {
+      sed -i "s/^[#%]*\(%global prerelease\).*$/\1 $prever/" "$X"
+      :
+    } || {
+      sed -i "/^\s*Name: / i %global prerelease $prever\n" "$X"
+    }
   }
 
   [[ -z "$ver" || "$prever" == '.' ]] \
@@ -306,7 +310,8 @@ xv="`rev <<< "$f" | cut -d'-' -f1 | rev`"
   :
 } || {
   prever=
-  sed -i "s/^[#%]*\(%global prerelease\).*$/#%\1 /" "$X"
+  [[ -n "$CON" ]] \
+    || sed -i "s/^[#%]*\(%global prerelease\).*$/#%\1 /" "$X"
 }
 echo
